ServiceWorkerGlobalScope is a global object and should be marked as [ImplicitThis] in the IDL https://bugs.webkit.org/show_bug.cgi?id=180615
Reviewed by Brady Eidson.

LayoutTests/imported/w3c:

Rebaseline WPT tests now that more checks are passing.

* web-platform-tests/service-workers/service-worker/interfaces-sw.https-expected.txt:
* web-platform-tests/workers/interfaces.worker-expected.txt:

Source/WebCore:

ServiceWorkerGlobalScope is a global object and should be marked as [ImplicitThis] in the IDL, similarly to what we do for Window.
This allows a getter to be fallback to the global object as ThisValue when the cast of the ThisValue to the expected type fails.

No new tests, rebaselined existing test.

* workers/DedicatedWorkerGlobalScope.idl:
* workers/WorkerGlobalScope.idl:
* workers/service/ServiceWorkerGlobalScope.idl:


Identifier: 196534@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@225718 268f45cc-cd09-0410-ab3c-d52691b4dbfc