[WinCairo] Enable running sharded tests https://bugs.webkit.org/show_bug.cgi?id=180660
Patch by Basuke Suzuki <Basuke.Suzuki@sony.com> on 2017-12-11
Reviewed by Alex Christensen.

The sharder splits the test inputs into groups of test based on the directory,
but separation of base name and directory name was wrong because of this bug.
Once sharded correctly, the execution of layout test on native Windows environment
is in parallel and much less time to run.

* Scripts/webkitpy/port/win.py:
(WinCairoPort):

Identifier: 196585@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@225770 268f45cc-cd09-0410-ab3c-d52691b4dbfc