Web Inspector: Slow open time enumerating system fonts (FontCache::systemFontFamilies) https://bugs.webkit.org/show_bug.cgi?id=180979 <rdar://problem/36146670>
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2018-01-03
Reviewed by Matt Baker.

Source/WebCore:

* platform/graphics/cocoa/FontCacheCoreText.cpp:
(fontNameIsSystemFont):
(WebCore::FontCache::systemFontFamilies):
Switch to the original Mac algorithm before r180979 that uses
CTFontManagerCopyAvailableFontFamilyNames. Previously this wasn't
available on iOS but now it is. This is a performance improvement on
both platforms, but significantly so on macOS. It also finds more,
valid, family names.

LayoutTests:

* inspector/css/get-system-fonts.html:
Cleanup the test a bit.

Identifier: 197094@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@226371 268f45cc-cd09-0410-ab3c-d52691b4dbfc